Resist more DOS keyword intern attacks
SNOOZE:URI-TO-ARGUMENTS is more careful when interning the keys in
key-value pairs found in the URI as keyword arguments.

It will only do so when a corresponding symbol with the same name is
found in the resource's package. This heuristic should prevent abusive
keyword interns.

* common.lisp (resource-package): New helper.
(uri-to-arguments): Be more careful when interning.
(read-for-resource): Use RESOURCE-PACKAGE

* snooze-tests.lisp (test-parse-uri): Explicitly intern BAR before
testing.