Close #36: allow users of MAKE-CLACK-APP to control response headers
They were already able to do this is the backend server was Hunchentoot,
but not if it was anything else.

This change dynamically binds a new *CLACK-RESPONSE-HEADERS* plist 
variable that request handlers can modify.  The resultwill then be part of 
the return value of the lambda returned by MAKE-CLACK-APP so Clack 
is able to make sense of it.